+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/playerimplem/PlayerImplemExoPlayer.kt
@@ -3,12 +3,9 @@ package com.github.florent37.assets_audio_player.playerimplem
 import android.content.Context
 import android.net.Uri
 import com.github.florent37.assets_audio_player.Player
-import com.google.android.exoplayer2.ExoPlaybackException
-import com.google.android.exoplayer2.ExoPlayer
-import com.google.android.exoplayer2.PlaybackParameters
+import com.google.android.exoplayer2.*
 import com.google.android.exoplayer2.Player.REPEAT_MODE_ALL
 import com.google.android.exoplayer2.Player.REPEAT_MODE_OFF
-import com.google.android.exoplayer2.SimpleExoPlayer
 import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory
 import com.google.android.exoplayer2.source.MediaSource
 import com.google.android.exoplayer2.source.ProgressiveMediaSource
@@ -23,17 +20,17 @@ class PlayerImplemExoPlayer(
         onBuffering: ((Boolean) -> Unit),
         onError: ((Throwable) -> Unit)
 ) : PlayerImplem(
-        onFinished=onFinished, 
-        onBuffering=onBuffering, 
-        onError=onError
+        onFinished = onFinished,
+        onBuffering = onBuffering,
+        onError = onError
 ) {
 
     private var mediaPlayer: ExoPlayer? = null
-    
+
     override var loopSingleAudio: Boolean
         get() = mediaPlayer?.repeatMode == REPEAT_MODE_ALL
         set(value) {
-            mediaPlayer?.repeatMode = if(value) REPEAT_MODE_ALL else REPEAT_MODE_OFF
+            mediaPlayer?.repeatMode = if (value) REPEAT_MODE_ALL else REPEAT_MODE_OFF
         }
 
     override val isPlaying: Boolean
@@ -53,34 +50,30 @@ class PlayerImplemExoPlayer(
         mediaPlayer?.playWhenReady = false
     }
 
-    fun getDataSource( context: Context,
-                               flutterAssets: FlutterPlugin.FlutterAssets,
-                               assetAudioPath: String?,
-                               audioType: String,
-                               networkHeaders: Map<*, *>?,
-                               assetAudioPackage: String?) : MediaSource {
+    fun getDataSource(context: Context,
+                      flutterAssets: FlutterPlugin.FlutterAssets,
+                      assetAudioPath: String?,
+                      audioType: String,
+                      networkHeaders: Map<*, *>?,
+                      assetAudioPackage: String?): MediaSource {
         try {
             mediaPlayer?.stop()
             if (audioType == Player.AUDIO_TYPE_NETWORK || audioType == Player.AUDIO_TYPE_LIVESTREAM) {
                 val uri = Uri.parse(assetAudioPath)
                 val userAgent = "assets_audio_player"
-                if(networkHeaders == null) {
-                    return ProgressiveMediaSource
-                            .Factory(DefaultDataSourceFactory(context, userAgent), DefaultExtractorsFactory())
-                            .createMediaSource(uri)
-                } else {
-                    return ProgressiveMediaSource.Factory(DataSource.Factory {
-                        val dataSource = DefaultHttpDataSource(userAgent)
-                        networkHeaders.forEach {
-                            it.key?.let { key ->
-                                it.value?.let { value ->
-                                    dataSource.setRequestProperty(key.toString(), value.toString())
-                                }
+
+                return ProgressiveMediaSource.Factory(DataSource.Factory {
+                    val allowCrossProtocol = true
+                    val dataSource = DefaultHttpDataSource(userAgent, DefaultHttpDataSource.DEFAULT_CONNECT_TIMEOUT_MILLIS, DefaultHttpDataSource.DEFAULT_READ_TIMEOUT_MILLIS, allowCrossProtocol, null)
+                    networkHeaders?.forEach {
+                        it.key?.let { key ->
+                            it.value?.let { value ->
+                                dataSource.setRequestProperty(key.toString(), value.toString())
                             }
                         }
-                        dataSource;
-                    }).createMediaSource(uri)
-                }
+                    }
+                    dataSource;
+                }).createMediaSource(uri)
             } else if (audioType == Player.AUDIO_TYPE_FILE) {
                 return ProgressiveMediaSource
                         .Factory(DefaultDataSourceFactory(context, "assets_audio_player"), DefaultExtractorsFactory())
@@ -104,6 +97,26 @@ class PlayerImplemExoPlayer(
         }
     }
 
+    private fun SimpleExoPlayer.Builder.incrementBufferSize(audioType: String): SimpleExoPlayer.Builder {
+        if (audioType == Player.AUDIO_TYPE_NETWORK || audioType == Player.AUDIO_TYPE_LIVESTREAM) {
+            /* Instantiate a DefaultLoadControl.Builder. */
+            val loadControlBuilder = DefaultLoadControl.Builder()
+
+/*How many milliseconds of media data to buffer at any time. */
+            val loadControlBufferMs = DefaultLoadControl.DEFAULT_MAX_BUFFER_MS /* This is 50000 milliseconds in ExoPlayer 2.9.6 */
+
+/* Configure the DefaultLoadControl to use the same value for */
+            loadControlBuilder.setBufferDurationsMs(
+                    loadControlBufferMs,
+                    loadControlBufferMs,
+                    DefaultLoadControl.DEFAULT_BUFFER_FOR_PLAYBACK_MS,
+                    DefaultLoadControl.DEFAULT_BUFFER_FOR_PLAYBACK_AFTER_REBUFFER_MS)
+
+            return this.setLoadControl(loadControlBuilder.createDefaultLoadControl())
+        }
+        return this
+    }
+
     override suspend fun open(
             context: Context,
             flutterAssets: FlutterPlugin.FlutterAssets,
@@ -115,23 +128,25 @@ class PlayerImplemExoPlayer(
         var onThisMediaReady = false
 
         try {
-            mediaPlayer = SimpleExoPlayer.Builder(context).build();
+            mediaPlayer = SimpleExoPlayer.Builder(context)
+                    .incrementBufferSize(audioType)
+                    .build()
 
             val mediaSource = getDataSource(
-                    context=context,
-                    flutterAssets=flutterAssets,
-                    assetAudioPath=assetAudioPath,
-                    audioType= audioType,
+                    context = context,
+                    flutterAssets = flutterAssets,
+                    assetAudioPath = assetAudioPath,
+                    audioType = audioType,
                     networkHeaders = networkHeaders,
-                    assetAudioPackage= assetAudioPackage
+                    assetAudioPackage = assetAudioPackage
             )
 
-            var lastState : Int? = null
+            var lastState: Int? = null
 
             this.mediaPlayer?.addListener(object : com.google.android.exoplayer2.Player.EventListener {
 
                 override fun onPlayerError(error: ExoPlaybackException) {
-                    if(!onThisMediaReady) {
+                    if (!onThisMediaReady) {
                         continuation.resumeWithException(error)
                     } else {
                         onError(error)
@@ -139,7 +154,7 @@ class PlayerImplemExoPlayer(
                 }
 
                 override fun onPlayerStateChanged(playWhenReady: Boolean, playbackState: Int) {
-                    if(lastState != playbackState) {
+                    if (lastState != playbackState) {
                         when (playbackState) {
                             ExoPlayer.STATE_ENDED -> {
                                 pause()
@@ -173,8 +188,8 @@ class PlayerImplemExoPlayer(
             })
 
             mediaPlayer?.prepare(mediaSource)
-        } catch (error: Throwable){
-            if(!onThisMediaReady) {
+        } catch (error: Throwable) {
+            if (!onThisMediaReady) {
                 continuation.resumeWithException(error)
             } else {
                 onError(error)